+++ src/main/kotlin/com/github/blarc/ai/commits/intellij/plugin/settings/AppSettings.kt
@@ -2,6 +2,7 @@ package com.github.blarc.ai.commits.intellij.plugin.settings
 
 import com.github.blarc.ai.commits.intellij.plugin.notifications.Notification
 import com.github.blarc.ai.commits.intellij.plugin.notifications.sendNotification
+import com.github.blarc.ai.commits.intellij.plugin.settings.prompt.Prompt
 import com.intellij.credentialStore.CredentialAttributes
 import com.intellij.credentialStore.Credentials
 import com.intellij.ide.passwordSafe.PasswordSafe
@@ -15,8 +16,8 @@ import com.intellij.util.xmlb.annotations.OptionTag
 import java.util.*
 
 @State(
-    name = AppSettings.SERVICE_NAME,
-    storages = [Storage("AICommit.xml")]
+        name = AppSettings.SERVICE_NAME,
+        storages = [Storage("AICommit.xml")]
 )
 class AppSettings : PersistentStateComponent<AppSettings> {
 
@@ -28,35 +29,8 @@ class AppSettings : PersistentStateComponent<AppSettings> {
     var requestSupport = true
     var lastVersion: String? = null
 
-    var currentPrompt: String = "basic"
-    var prompts: MutableMap<String, String> = mutableMapOf(
-        // Generate UUIDs for game objects in Mine.py and call the function in start_game().
-        "basic" to "Write an insightful but concise Git commit message in a complete sentence in present tense for the " +
-                "following diff without prefacing it with anything, the response must be in the language {locale} and must" +
-                "not be longer than 74 characters. The sent text will be the differences between files, where deleted lines" +
-                " are prefixed with a single minus sign and added lines are prefixed with a single plus sign.\n" +
-                "{diff}",
-        // feat: generate unique UUIDs for game objects on Mine game start
-        "conventional" to "Write a clean and comprehensive commit message in the conventional commit convention. " +
-                "I'll send you an output of 'git diff --staged' command, and you convert " +
-                "it into a commit message. " +
-                "Do NOT preface the commit with anything. " +
-                "Do NOT add any descriptions to the commit, only commit message. " +
-                "Use the present tense. " +
-                "Lines must not be longer than 74 characters. " +
-                "Use {locale} language to answer.\n" +
-                "{diff}",
-        // ✨ feat(mine): Generate objects UUIDs and start team timers on game start
-        "emoji" to "Write a clean and comprehensive commit messages in the conventional commit convention. " +
-                "I'll send you an output of 'git diff --staged' command, and you convert " +
-                "it into a commit message. " +
-                "Use GitMoji convention to preface the commit. " +
-                "Do NOT add any descriptions to the commit, only commit message. " +
-                "Use the present tense. " +
-                "Lines must not be longer than 74 characters. " +
-                "Use {locale} language to answer.\n" +
-                "{diff}",
-    )
+    var prompts: MutableMap<String, Prompt> = initPrompts()
+    var currentPrompt: Prompt = prompts["basic"]!!
 
     companion object {
         const val SERVICE_NAME = "com.github.blarc.ai.commits.intellij.plugin.settings.AppSettings"
@@ -64,9 +38,16 @@ class AppSettings : PersistentStateComponent<AppSettings> {
             get() = ApplicationManager.getApplication().getService(AppSettings::class.java)
     }
 
-    fun getPrompt(diff: String) = prompts.getOrDefault(currentPrompt, prompts["basic"]!!)
-        .replace("{locale}", locale.displayName)
-        .replace("{diff}", diff)
+    fun getPrompt(diff: String): String {
+        val content = currentPrompt.content
+        content.replace("{locale}", locale.displayName)
+
+        return if (content.contains("{diff}")) {
+            content.replace("{diff}", diff)
+        } else {
+            "$content\n$diff"
+        }
+    }
 
     fun saveOpenAIToken(token: String) {
         try {
@@ -84,10 +65,10 @@ class AppSettings : PersistentStateComponent<AppSettings> {
 
     private fun getCredentialAttributes(title: String): CredentialAttributes {
         return CredentialAttributes(
-            title,
-            null,
-            this.javaClass,
-            false
+                title,
+                null,
+                this.javaClass,
+                false
         )
     }
 
@@ -104,6 +85,44 @@ class AppSettings : PersistentStateComponent<AppSettings> {
         }
     }
 
+    private fun initPrompts() = mutableMapOf(
+            // Generate UUIDs for game objects in Mine.py and call the function in start_game().
+            "basic" to Prompt("Basic",
+                    "Basic prompt that generates a decent commit message.",
+                    "Write an insightful but concise Git commit message in a complete sentence in present tense for the " +
+                            "following diff without prefacing it with anything, the response must be in the language {locale} and must " +
+                            "NOT be longer than 74 characters. The sent text will be the differences between files, where deleted lines" +
+                            " are prefixed with a single minus sign and added lines are prefixed with a single plus sign.\n" +
+                            "{diff}",
+                    false),
+            // feat: generate unique UUIDs for game objects on Mine game start
+            "conventional" to Prompt("Conventional",
+                    "Prompt for commit message in the conventional commit convention.",
+                    "Write a clean and comprehensive commit message in the conventional commit convention. " +
+                            "I'll send you an output of 'git diff --staged' command, and you convert " +
+                            "it into a commit message. " +
+                            "Do NOT preface the commit with anything. " +
+                            "Do NOT add any descriptions to the commit, only commit message. " +
+                            "Use the present tense. " +
+                            "Lines must not be longer than 74 characters. " +
+                            "Use {locale} language to answer.\n" +
+                            "{diff}",
+                    false),
+            // ✨ feat(mine): Generate objects UUIDs and start team timers on game start
+            "emoji" to Prompt("Emoji",
+                    "Prompt for commit message in the conventional commit convention with GitMoji convention.",
+                    "Write a clean and comprehensive commit message in the conventional commit convention. " +
+                            "I'll send you an output of 'git diff --staged' command, and you convert " +
+                            "it into a commit message. " +
+                            "Use GitMoji convention to preface the commit. " +
+                            "Do NOT add any descriptions to the commit, only commit message. " +
+                            "Use the present tense. " +
+                            "Lines must not be longer than 74 characters. " +
+                            "Use {locale} language to answer.\n" +
+                            "{diff}",
+                    false)
+    )
+
     class LocaleConverter : Converter<Locale>() {
         override fun toString(value: Locale): String? {
             return value.toLanguageTag()
@@ -113,5 +132,4 @@ class AppSettings : PersistentStateComponent<AppSettings> {
             return Locale.forLanguageTag(value)
         }
     }
-
 }